moved user
Moved user class into page3
and updated the path
